refactor(timeline): extract visual frame geometry into useTimelineItemBounds
Move the leftFrame/rightFrame computation, the contentPreviewItem
derivation chain (trim/roll/slide/slip with source-frame adjustments),
and the visualLeftFrame/visualWidthFrames calculation into a single
hook that takes the preview shifts and trim/stretch state.

Source-frame conversions, slide-continuity math, and isCompactWidth
calculation all live in the hook now.
